Two changes:
 * Cython wrapper now lives in python/ and compiles
 * Changes a_value to FLOAT and TimeFieldLastUpdated to double as well

+++ src/clib/chemistry_data.h
@@ -55,7 +55,7 @@ typedef struct
   int TabulatedLWBackground;
   float RadiationFieldRedshift;
 
-  FLOAT TimeFieldLastUpdated;
+  double TimeFieldLastUpdated;
   int  RadiationShield;
 
   /* Frequency dependant information (constant). */
